# PERF: RangeIndex.get_loc