@jorisvandenbossche 2020-01-13
How does this change affect `__contains__` ?