@jorisvandenbossche 2020-01-13
The timings you showed were then also not relevant?